Move postgres utility code to a new file
